Whether m17n mode is used should depend on the .conf file, not the language
Not all languages other then English necessarily use m17n
transliteration. For example German will usually be typed with a
German keyboard layout. Using m17n transliteration like latn-post or
latn-pre for German is possible, but not so common. So for some
languages both is possible, using m17n transliteration or not using
it. Therefore, it is better to make m17n mode depend on
the respective option in the .conf file.