Make sure word is correctly encoded
The parameter “word” of words_start() should always contain
Unicode. But it contained UTF-8 because it was accidentally converted
back to UTF-8 from Unicode in the select_words() function, this was a
side effect of using str(). I prevented the conversion to UTF-8 now by
using eval(repr()) instead of str().  Nevertheless I left the check
whether “word” contains Unicode in words_start(), it is not
needed anymore but does not hurt there.