Let first letter start with index 1 in autogenerated tabdict
The hardcoded tabdict starts with the first letter in index 1
and has a dummy entry in index 0:

    tab_dict = {
        '0':0,
        'a':1, 'b':2, 'c':3, 'd':4, 'e':5,
    ...

But index 0 is ignored in another_lang_deparser(id),
i.e. another_lang_deparser(0) just returns '':

    def another_lang_deparser(id):
        if id:
            id = int(id)
            if id in id_tab_dict:
                return id_tab_dict[id]
        else:
            return ''

To make the autogenerated tabdict work the same way as
the hardcoded one, add the dummy entry in index 0 and
start with the first letter in index 1 as well.