Add translation of X11 keysyms to Unicode
This is needed for example to make the Russian native keyboard work.

For example, the keysym for ф U+0444 CYRILLIC SMALL LETTER EF could
be 0x1000444 or 0x6c6 which corresponds the keysym name
“Cyrillic_ef”.  But unichr(0x1000444) would fail with “ValueError:
unichr() arg not in range(0x110000)” because Unicode code points
cannot be larger then 21 bit. And unichr(0x6c6) is ARABIC LETTER OE.
To be able to convert these keysyms to Unicode correctly, the function
keysym2unichr() is added here. Both keysym2unichr(0x1000444) and
keysym2unichr(0x6c6) return the ф.