Make sure the input of transliterate() is UTF-8 encoded
The c-function called in transliterate(self, _input) needs _input in
UTF-8 encoding. If this function gets _input in Python’s Unicode
type, accept that as well by converting it to UTF-8 first.

This makes it also possible to re-apply

    commit 3aae789e46e143722d2e9c081544f2c543f16061
    Author: Mike FABIAN <mfabian@redhat.com>
    Date:   Fri Oct 12 10:09:46 2012 +0200

        Revert part of commit “Make sure word is correctly encoded” to fix Marathi

        commit f277f35a50eb0c39f0a4a5864e91fb8ef848488a
        Author: Mike FABIAN <mfabian@redhat.com>
        Date:   Fri Oct 5 08:48:05 2012 +0200

            Make sure word is correctly encoded

        broke it for Marathi. Revert the part which broke it.

The reason it was broken for Marathi was transliterate
got Unicode input instead of UTF-8 which it could not handle.
But now it can handle both so this commit can be applied again
without breaking Marathi.