use “layout = default” instead of “layout = us” in all .conf files
/usr/share/ibus/component/typing-booster.xml contains

    <engines exec = '/usr/libexec/ibus-engine-typing-booster --xml'/>

and this produces output like:

    ...
    <engine>
        <name>typing-booster:ru_RU</name>
        <longname>Russian - RU(Hunspell)</longname>
        <language>ru</language>
        <license>GPL</license>
        <author>Mike FABIAN &lt;mfabian@redhat.com&gt;</author>
        <icon>/usr/share/ibus-typing-booster/icons/english.png</icon>
        <layout>us</layout>
        <description>This is a Russian typing booster engine table for ru_RU language
</description>
        <page_size>6</page_size>
        <symbol>ru_RU(Hunspell)</symbol>
        <setup>/usr/libexec/ibus-setup-typing-booster</setup>
    </engine>
    ...

The content between the <layout>...</layout> tags comes from the .conf
files. If “layout = us” is used, the keyboard layout is switched to
US English layout when that ibus-typing-booster engine is activated.
For the Russian typing booster using the native Russian keyboard layout
this makes no sense. But if one used “layout = ru” there, then
the Russian typing booster would not work for the other imes it supports.
Currently, ru_RU.conf contains:

    imes = Native Keyboard:NoIme,Simulated Russian Keyboard:ru-kbd,Transliteration:ru-translit,Phonetic:ru-phonetic,Yawerty:ru-yawerty

For “Native Keyboard:NoIme”, a Russian keyboard layout is needed.
For all the others, any keyboard layout where ASCII can be typed
can be used.

At the moment I have no better idea than to use “layout = default”
in the .conf files because of this problem. When “layout = default”,
ibus does not change the keyboard layout when activating an input
method engine, i.e. it leaves the decision which keyboard layout to
use to the user. A user who prefers “Native Keyboard:NoIme” for
the Russian typing booster would need to make sure to set the Russian
keyboard layout, a user who prefers one of the m17n transliteration
input methods for the Russian typing booster would need to make sure
*not* to set the Russian keyboard layout but any other layout which
can be used to type ASCII.