Make German input typed in NFD work
NFD input, i.e. base character followed by combining character, works
as well with this patch. The NFD input is converted to NFC. Matching
in the dictionary and user databases is then done in NFC only.  A
consequence is that the result is later committed as NFC as well. This
makes it impossible to input NFD while using the typing booster
because the result is automatically normalized to NFC. But this should
be no problem, usually the normalization is wanted. If one really
wants to type something in NFD, one has to switch the typing booster
off for that.

There is more to do to make Korean work because ko_KR.dic is in NFD.