Fix the problem that the user defined phrases are lost when switching engines
When defining a new phrase it was lost after switching engines.

How to reproduce:

- Select British English typing booster
- type “supercalifragilisticexpialidocious” followed by space
- type “super” again and confirm that “supercalifragilisticexpialidocious”
  is shown as the first suggestion in blue, not black, because this
  is user defined.
- switch to any other ibus engine, for example Danish typing booster
- switch back to the British English typing booster
- type “super” again, the suggestion “supercalifragilisticexpialidocious”
  is not shown anymore, it is lost in memory and it has not been saved
  to ~/.local/share/.ibus/hunspell-tables/typing-booster\:en_GB-user.db
  either (saving to the .db files happens only when ibus is restarted).

The reason is that in create_engine() in factory.py, a new
db is created always. But a db for an engine with that name
might already exist from using the same engine previously, it
should be reused not discarded.