Split only at the first = in a line in a .conf file
That means, attr may not contain a '=' but val may contain '='
allowing val to contain a '=' might be useful for the values of the
attributes lang_chars and valid_input_chars.  Even if '=' is not
really useful there, it makes the parsing more robust if a '=' ends up
accidentally there by trying to figure out lang_chars and
valid_input_chars semi-automatically using a script.