Do not forget the input method used last when activating a previously used engine
Similar to https://bugzilla.redhat.com/show_bug.cgi?id=865671
ibus-typing-booster forgot the input method used last with
one of its engines.

Steps to reproduce:

- Activate ibus-typing-booster for German.
  The default input method for German is to use the native
  keyboard layout, i.e. to type an ä one needs to input this
  directly via the keyboard, typing a" (a followed by double quotes)
  just inserts a" and does nothing special.
- Now select the “Latin-Post” input method from the
  panel.
- Type a" again, it gets transliterated to ä now and words
  starting with ä show up as suggestions
- Switch to any other ibus input engine, for example an
  ibus-typing-booster engine for a different language, e.g.
  ibus-typing-booster for English. Or any other input engine
  like ibus-anthy for Japanese.
- Switch back to ibus-typing-booster for German.
- Type a" again, it inserts just a" and is not transliterated
  to ä anymore, i.e. the input method selected before when
  using the German ibus-typing-booster last has been forgotten.

Fix:

When an engine is created in ibus-typing-booster, do not throw away
the previous engine and create a new engine from scratch each
time. Rather keep all previously used engines in a hash table and
reuse them when a previously used engine is activated again.
Reusing the previous instance of an engine keeps the internal state
of that engine. This way, the input method selected when using that
engine is not forgotten.