Show exact matches in the .dic files as suggestions as well
Until now, the regular expression used for matches in the .dic files
required a word in the dictionary to match the characters already
typed plus at least one more character from the characters listed in
the lang_chars option. I.e. if a word which is in the dictionary
was typed completely, no suggestion was shown anymore.

This had 2 disadvantages:

1) it was inconsistent with how suggestions for user defined words are
   shown. For user defined words, i.e. words saved by the user in
   a database  like
   ~/.local/share/.ibus/hunspell-tables/typing-booster\:de_DE-user.db
   a suggestion is shown even for exact matches. It makes more sense
   if suggestions for words from the global dictionaries behave the
   same way.
2) If a word is typed completely and no suggestion is shown, the
   user cannot distinguish whether the word has been typed correctly
   and matches a word in the global dictionary or whether the last
   letter has been mistyped. It is better to show the suggestion for
   an exact match as well to be able to see whether the typed word matches
   or not. If it does not match and the user commits it by typing a space,
   the word is added as a user defined word to the user dictionary.
   This should be clearly visible.

Therefore, this commit changes the regular expression a bit to match
also when a typed word matches a word in the global dictionary
exactly.