Make enabling the lookup table with the TAB key work correctly.
Resolves: #869687 https://bugzilla.redhat.com/show_bug.cgi?id=869687

When the option to use the TAB key to enable suggestions was set, the
lookup table did not show up immediately after pressing TAB but only
after typing the next character. And it immediately disappeared again
after typing one more key. It also disappeared again when the next key
typed was arrow-up or arrow-down, making it impossible to select a
candidate using arrow-up or arrow-down.

I fixed this and added the following behaviour:

When the option to use the TAB key to enable suggestions is set, I
made the TAB key toggle showing the lookup table, i.e. one TAB enables
the lookup table, the next TAB disables it again.  On top of that, the
mode is reset to hide the lookup table after each commit and after
deleting the complete preedit string for example using backspace.

This could  be done differently, for example one could easily
make typing TAB once enable the lookup table permanently and
disable it permanently after the next TAB. Let’s see what turns out
to be more useful.