Simplify code in add_input()
Treat the cases with and without m17n transliteration more similar.
The code becomes shorter and simpler.

This should also make it easier to modify the entered text for example
to do normalization to NFC or NFD. Adding the entered characters
individually to self._tabkey_list as it was done before in the case
where m17n transliteration was not used makes something like
normalization harder because normalization would need to be applied
to self._tabkey_list as well then.

Rename self._latin_chars to self._typed_chars because this variable
does not necessarily contain Latin. It can just as well contain
Cyrillic for example when typing Russian with a native keyboard
layout.