Make the setup tool find the right config file in gnome-shell on Fedora 18
On Fedora 17, ibus-setup calls the setup tool with the options
as specified in the output of

    “/usr/libexec/ibus-engine-typing-booster --xml”

I.e. if this output contains
    <engine>
        <name>typing-booster:es_ES</name>
        ...
        <setup>/usr/libexec/ibus-setup-typing-booster --config-file es_ES.conf</setup>
    </engine>

then the setup tool is called with exactly that command line.

However, when using gnome-shell on Fedora 18, the setup tool is called
like this:

    “IBUS_ENGINE_NAME=typing-booster:es_ES /usr/libexec/ibus-setup-typing-booster”

Therefore, figure out the config file name from the engine name if
the command line option is not given but the environment variable
IBUS_ENGINE_NAME is set.