pass the name of the affix file when creating the Hunspell class object
The variable self.aff_name in “class Hunspell” was always set to the
fallback 'en_US.aff' because the parameter aff_name in the constructor
was not passed. This didn’t really matter so far as the affix file
was not really used anyway. To improve the suggestions, it makes sense
to be able to use the correct affix file as well, therefore this
parameter should be passed correctly.