Make ap_candidate() more readable by removing some unused code
The code removed (which was already commented out) was
inherited from ibus-table and used to display keys which
would complete a candidate when typed in different colour.
I.e. when using for example wubi-jiidian86, typing “a”
would display a lookup table like

    工
    a
    -------
    1.| 工
    2.| 区q
    3.| 东i
    4.| 世n
    5.| 或k
    6.| 式a

where the Latin letters behind the candidates 2.-6. are displayed in a
different colour. This is not needed for ibus-typing-booster.