Display of the number of candidates needs to be updated on page-up and page-down
Updating the lookup table only on page-up and page-down is not enough,
the auxiliary text needs to be updated as well because the number
of candidates may be displayed there. I use self._update_ui() because
this does:

        self._update_lookup_table ()
        self._update_preedit ()
        self._update_aux ()

updating the preedit text is actually unnessesary in case of
ibus-typing-booster. For ibus-typing-booster, the preedit text does
not change when moving through the lookup table. For ibus-table, the
preedit text does change though when moving through the lookup table
and currently ibus-table has a bug there, it does not update the
preedit text on page-up and page-down although it should.