Get rid of tab_dict
Do not convert the input characters to integers to store in the database,
i.e. do not use database rows like:

1|4|4|51|62|89|58|||||||||||||||||||||||||||||||||||||||||||||||grün|0|20

Instead store input characters directly in the database, i.e.
now the database rows look like this:

3|4|4|g|r|ü|n|||||||||||||||||||||||||||||||||||||||||||||||grün|0|3

The “tab_dict” was used for the conversion between the characters
and the integers. Therefore, it is not necessary anymore now.  And a
lot of code doing the conversion could be removed. The code is shorter
and easier to read now.