Fix optimize_database() function to not fail when called.
This function was never used, so maybe I could have deleted it as
well.  It is used in tabcreatedb.py in ibus-table though, so maybe it
has some useful effect. So for the moment I just removed the stuff
inherited from ibus-table to make it not fail when called and call it
on startup. Let’s see whether this does something useful or not.  If
it does not do anything useful now, we could maybe improve it.  If it
turns out to be too slow to always call it at startup, we could call
it optionally. Or remove it completely later when it turns out that
there is no benefit.