Use “Language - IN (Hunspell)” as ime_name for the Indian languages
Some of the Indian languages still had “Language - Inscript (Hunspell)”
as ime_name. But this is a bit misleading because Inscript is only
one of the available input methods in the setup. Therefore,
use “Language - IN (Hunspell)” where IN is the country code to make it
the same format as for the other languages.