Don’t call self.optimize_database()
self.optimize_database() is currently called in tabsqlitedb which is used by

    python engine/main.py --xml

to list the engines. So it should not produce extra output like the
debug message I removed because it makes the xml output invalid.
Even without the debug message, it makes the listing of the engines
which is already slow even slower.

I am not sure what to do with this function Maybe it can be improved
to do something useful and be called optionally later.  For the moment
I just comment it out.