Speed up generating the xml list of the engines
Generating the xml list of the engines by calling “python
/usr/share/ibus-typing-booster/engine/main.py --xml” was extremely
slow because a tabsqlite object was created for each config file only
to be able to parse the config file.  Because the config file parser
was part of the class “tabsqlite”.

The main reason why creating a tabsqlite object is so expensive is
that it triggers loading the hunspell dictionary into RAM. Loading
more then 70 big hunspell dictionaries is of course slow.

But the dictionaries are not needed at all for creating the xml, then
are only needed when the engine is really used. So one way to speed
this up a lot would have been to delay loading of the hunspell
dictionary to when it is really needed.  I tried this and it really
reduces the time for creating the xml from 7.5 seconds to 0.4
seconds. But even if the loading of the hunspell dictionary is
delayed, creating a tabsqlite object still does a lot of stuff which
is not needed at all for producing the xml, for example it opens the
sqlite user databases.

Therefore, I restructured the code a bit to put the config file
parsing in a new class which can then be used by the “tabsqlite”
and can also be used directly when creating the xml.

Time for “python /usr/share/ibus-typing-booster/engine/main.py --xml”

before: 7.5 sec
now: 0.13 sec