Remove the options “user_can_define_phrase” and “dynamic_adjust” from code and .conf files
We always want phrases to be dynamically adjusted. For
ibus-typing-booster it makes no sense not to learn fromm the user
input and adjust the frequency of the phrases. We also want to allow
the user always to define new phrases, only because a word is not
suggested by hunspell it does not mean that we should not put it
into the database and count the frequency. Words typed often
by the user should always get their frequency data updated in the
user database, no matter whether that word is in the hunspell
dictionary or not.