Merge candidates which have the same resulting phrase in select_words()
Merge all the candidates which share the same phrase into one, summing
up their user frequencies.

For example, if the database contains the following rows:

1|5|6|colou|colour|0|1
2|3|6|col|colour|0|2
3|2|6|co|colour|0|1
4|2|4|co|cold|0|1
5|9|10|conspirac|conspiracy|0|1
6|6|10|conspi|conspiracy|0|1
7|1|10|c|conspiracy|-1|1

and the current input_phrase is “co”, the last line with
input_phrase = “c” and phrase “conspiracy” would not have matched
here because we used “LIKE "co%"” in the SQL statement. It we used
the remaining candidates as is, we would see duplicate entries in the
lookup table, for exaple we would see “colour” 3 times in the lookup
table which makes no sense. Therefore, we merge the candidates which
share the same phrase and sum the frequencies.  Doing this we get for
the above example:

1|2|6|co|colour|0|4
2|2|4|co|cold|0|1
3|9|10|co|conspiracy|0|2

I.e. for the input phrase pattern “co%”, “colour” matched 4 times
total, “conspiracy” matched 2 times total and “cold” matched just
once.