Remove phrases always from the user database when typing Alt+Number
It should not matter whether the phrase selected with Alt+Number is a
system phrase or user defined phrase, in both cases it should be
removed from the user database to score down the likelyhood of this
phrase being suggested again.

The user should not need to think about whether this is a system
phrase or a user defined phrase, selecting a phrase with Alt+Number
just means: “I do not want this to be suggested here, reduce the
score of this”.

Also, make the function remove_phrase() use keyword arguments
for input_phrase and phrase instead of passing a complete
database row where most of the columns are ignored anyway,
this improves readability a bit, I think.