simplify sync_usrdb() and make code more readable
Instead of

    SELECT * FROM mudb.phrases;

use

    SELECT input_phrase, phrase, freq, user_freq FROM mudb.phrases;

This makes the code of sync_usrdb() understandable without knowing
and counting the column numbers of the database.
Also makes it easier to add or remove columns when they are not
used here anyway. I think of removing “mlen” and “clen” soon,
after this change sync_usrdb() is not affected by removing
“mlen” and “clen” anymore.