Return lists of (phrase, user_freq) pairs only in select_words()
Do not return lists of whole database rows in select_words() if most
of the columns are not used at all in hunspell_table.py.  It is
useless to return “id”. “input_phrase” is already known and never
changes, so that is useless in the return value as well.  “freq” is
quite pointless as well, it is not helpful to distinguish between
“user defined phrases” and “system phrases” in the lookup table.
Deleting/scoring down phrases which appear in the lookup table by
typing Alt+Number now works for both “user defined phrases” and
“system phrases”, using different colours for them in the
lookuptable just makes it more confusing without adding any real
value.