Do not use lang_chars for matching in the hunspell dictionaries, return immediately if input contains a “/”
Resolves: #966947 https://bugzilla.redhat.com/show_bug.cgi?id=966947

We matched a word in the dictionary by trying to match the word as
typed so far followed by characters allowed in the language,
i.e. followed by characters listed in the value of “lang_chars” in
the .conf file.

But it is much easier to just match for the word followed by anything
until a “/” or the end of a line is encountered.  “/” is the only
character in hunspell dictionaries which can end a word before the end
of the line is reached because “/” seperates words from affix rules.

Using the characters allowed in the language has no value here
and makes it error prone if the list of allowed characters
in the .conf file is incomplete.

If the input already contains a “/”, no match is possible anymore,
we should return an empty list immediately.

We had a bug here, for example the da_DK.dic contains stuff like

    a/ACJSTVW

and typing “a/” as the input matched “a/ACJSTVW” although
“ACJSTVW” are just affix rules.

See: https://bugzilla.redhat.com/show_bug.cgi?id=966947