Remove lang_chars variable
lang_chars was originally used to genereate the language specific
tab_dict. But tab_dict is gone now.

The last use of lang_chars was for matching in the hunspell
dictionaries, but it was not a good idea to use lang_chars there
anyway and the matching has been changed to match everything up to a
“/” which seperates words from affixes.

So lang_chars is completely useless now and can be removed.