Get rid of the the variable “valid_input_chars”
Resolves: rhbz#968209 https://bugzilla.redhat.com/show_bug.cgi?id=968209

Listing all valid input characters for a language is tedious
and error prone.

Even if the list of valid input characters in the config file
is correct for a language, it is still not nice to prevent typing
other characters.

For example, if I am using ibus-typing-booster to type German, I still
may want to type some non-German words like “São Paulo”, it is not
nice if this fails only because “ã” happens to be not listed as a
valid input character for German (Currently typing “São” with the
German ibus-typing-booster gives the result “ãSo”).

Allowing all characters as input except those which are needed to
navigate the lookup table and select candidates avoids this problem.

And typing something like “São” then also correctly adds it to the
user database of phrase frequencies so a German user who types that
often will even get it suggested.