Remove variables max_key_length and auto_commit
Since the database format is changed to have the input phrase in a
single column of type TEXT instead of storing each character of the
input phrase in a seperate column, the maximum length of an input
phrase has become huge (maximum string length in sqlite3, by default
10^9, see http://www.sqlite.org/limits.html).

There is no need to have an artificially low limit in the config
files now, this was only caused by the old database design.

In the code, self._max_key_length was only used together with
auto_commit, if auto_commit was true and the length of the input had
reached t self._max_key_length, it was committed. That doesn’t
seem to make much sense either, especially because auto_commit
was *always* False.