Remove unreachable code
editor.add_input() *always* returns the return value of
editor.update_candidates() which *always* returns “True”.  (It was
always like this, even in the initial commit of ibus-typing-booster,
only in ibus-table it is possible that editor.update_candidates()
returns something different than “True”).

So this code could never be reached.