Fix the problem that after “page down” the first “arrow down” does not move down in the lookup table
There were even two variables to remember whether navigation in the
lookup table had started or not: self._first and self.is_down_press.
These were used somewhat incorrectly, for example page_down(),
page_up(), arrow_up() set them to indicate that the navigation of the
lookup table had started. Therefore, the first arrow_down() after one
of these other lookup table navigation failed to work.

But the intention of this was only that when the navigation of the
lookup table was started with arrow_down(), it should
only make the cursor in the top row of the lookup table visible
and not move the cursor down to the second row already.

To achieve this, no extra variables are necessary, it seems more
natural to use “self._lookup_table.cursor_visible” to check whether
navigation of the lookup table has started or not.