Replace unencodable characters when asking pyhunspell for suggestions
Use 'replace' when encoding the word into the dictionary encoding to
ask pyhunspell for suggestions. For example, when only one character
is in the input which is not encodable into the dictionary encoding,
this character maybe a typo and one may still get a useful match from
pyhunspell if one queries using a replacement character.  So asking
pyhunspell using replacement characters is better than skipping the
query completely.