When committing by tying RETURN or ENTER, update frequency data in user database
When committing by typing RETURN or ENTER, the frequency data in the
user database was not updated. This commit fixes this by merging the
case for IBus.KEY_Return into IBus.KEY_KP_Enter into the case for
IBus.KEY_space, changing this case slightly not to add u' ' to the
commit string but rather passing the SPACE key through after
committing.