Do not try to match very long words in the hunspell dictionaries
If the input phrase is very long, don’t try looking
something up in the hunspell dictionaries. The regexp match
gets very slow if the input phrase is very long. And there
are no very long words in the hunspell dictionaries anyway,
the longest word in the German hunspell dictionary currently
seems to be “Geschwindigkeitsübertretungsverfahren” trying
to match words longer than that just wastes time.

When trying to match words no matter how long they are,
ibus-typing-booster cannot keep up with the input speed when holding
down a key permanently and using the keyboard autorepeat. After around
50 characters have been inserted, it becomes very slow and then
suddenly characters don’t get added to preëdit anymore but are passed
through to the application. It is probably not easy to run
into that problem with “normal” typing, but it might happen
when trying to type very long strings.

By limiting the length of words where a match in the hunspell
dictionaries is still attempted, ibus-typing-booster can still
keep up even with fast keyboard autorepeat.