Get rid of “freq” column in databases
This column is completely useless for improving the prediction
quality.  We do not really care whether a phrase is suggested because
of hunspell or because the user has typed it recently. If the user has
typed it recently, it has user_freq >= 1. If it is only suggested by
hunspell and has never been typed by the user, we give it user_freq =
0. Having an extra column in the database just to remember whether
this phrase is in the hunspell dictionary does not add any value at
all.