Remove too simpleminded auto-capitalization
If the user had nothing selected in the lookup-table, i.e.
not yet used the arrow up/down or page up/down keys to navigate
the lookup table and typed RETURN, ENTER, or SPACE, then
the first word in the lookup table was committed instead of
the user input if that word compared case insensitively equal
to the user input.

I.e. if the user types

   japanese

and the first word in the lookup table happens to be

   Japanese

then “Japanese” is committed. In this case, this would most likely
be correct because “Japanese” is spelled correctly with a capital
“J”.  But in most cases this just causes random capitalization
errors for me. Many German words exists with and without
a capital first letter, which one is correct depends on context.
For example, in German “Wagen” is a noun and means car whereas
“wagen” is a verb which means “to dare”.

With this simple auto-capitalization, it is not possible to commit the
word exactly as typed at all if the first candidate in the lookup
table happens to compare case insensitively equal to the typed word,
the candidate from the lookup table is always committed, never the
word typed by the user, there is no override.

So I remove this automatic capitalization for the moment, it causes
more errors than it fixes.

Maybe I introduce automatic capitalization again later with a more
intelligent way to do it. But even then it should be an option which
can be switched off.