Create a VIEW for “LIKE input_phrase%” in select_words() and use that in the following SELECT statements
Instead of using “LIKE input_phrase%” again in all 6 SELECT
statements, do it only once by creating a VIEW first and using that in
the select statements.  Makes select_words() more than 10 times
faster:

Profiling results  with old code:

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
       47    0.015    0.000   63.304    1.347 tabsqlitedb.py:367(select_words)

Profiling results with new code:

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
       78    0.026    0.000    8.128    0.104 tabsqlitedb.py:367(select_words)

In both cases the same long sentence was typed. The sentence appears
in a long book which was used to train the database (The database had
around 300000 records after training). The speed difference is already
obvious when typing even without measuring.

I wanted to do this with a VIEW already when I first implemented this,
I only failed because sqlite does not allow to create view with
parameters:

    “OperationalError: parameters are not allowed in views”

And here I wanted to create a view using the user input “input_phrase”
as a parameter.

A workaround for this problem is not to pass “input_phrase” as a
parameter but directly assemble the SQL statement to create the view
with Python string operations including “input_phrase”. Assembling
SQL statements like this is not recommended because of the possibility
of SQL injection attacks. But in this case there is no other choice
because using a parameter does not work. And if the user input
“input_phrase” is properly quoted when assembling the SQL statement
(Remove all NUL characters, replace " with "" and wrap the whole
string in double quotes) then it should be OK.