Ignore most punctuation characters and mathematical symbols when tokenizing
Resolves: rhbz#979939 - It will be good to remove punctuation marks and quote from words before adding to Database.
See: https://bugzilla.redhat.com/show_bug.cgi?id=979939

When reading a file to train the database, remove leading and trailing
punctuation and mathematical symbol characters.

Do the same when learning tokens from input the user types.

When the user starts typing something with a punctuation character,
for example when the user types

    “col

strip the leading “ to make it possible to match 'colour' in the
database.  Because such punctuation characters are not added to the
database anymore, “col can never match '“colour' in the database
now. In order to match something useful nevertheless when the user
input starts with such a punctuation characters, strip the leading
punctuation characters before doing the database matching and add them
again when populating the lookup table.