Add an option to try completion only when a minimum number of characters has been typed
Some users reported that they find it distracting when the lookup
table pops up all the time and requested an option to try completion
only if a certain number of character has been typed already. I.e. use
completion only for longer words. Short strings are still in preëdit
until committed and added to the database, but no lookup table pops up
anymore when the number of characters typed is still below the limit.

For very fast typists, this may be better then popping up the lookup
table after the first character typed.

Of course, to save as many keystrokes as possible, completion should
be tried as early as possible. Therefore, the default is 1,
i.e. completion is attempted after the first character typed.