Use normalization form NFD internally for Korean as well
Use the normalization forms as follows:

NFD:  for internal matching in the sql database and grepping
      in the hunspell dictionary for all languages except
      Korean
NFC:  for display in the preëdit and the lookup table, for calculating
      the caret position in the preëdit, for commiting.
      Also for input into pyhunspell.

I used NFKD for Korean internally for the database.  But that makes it
difficult to use the same database for all languages.

And the reason that I used NFKD for Korean is only that the 2
Korean input methods we have in m17n produce Jamo  from
the Hangul compatibility block and not from the Hangul Jamo area.
For example transliterationg 'h' using /usr/share/m17n/ko-romaja.mim
gives

U+314E ㅎ

not

U+1112 ᄒ

https://en.wikipedia.org/wiki/List_of_hangul_jamo

To match correctly in the hunspell dictionary
/usr/share/myspell/ko_KR.dic we need the latter (U1112 from the Hangul
Jamo area).

But instead of converting to NFKD when writing to our database and
when matching in the hunspell dictionary, we can just as well convert
the transliteration output immediately to NFKD.  Converting the
transliteration output of /usr/share/m17n/ko-romaja.mim and
/usr/share/m17n/ko-han2.mim to NFKD converts U+314E ㅎ to U+1112 ᄒ
etc.

Further converting these Jamo from the Hangul Jamo area (like U+1112 ᄒ)
to any of the normalization forms NFC, NFKC, NFD, and NFKD does not
change them anymore. I.e. when the conversion to NFKD is done right
after the transliteration when using one of these 2 Korean input methods,
the matching works as well and we do not have to use a different
normalization form for Korean in our database.