Use a single user database for all engines
To make ibus-typing-booster more multilingual, it is useful to add
trigrams always to the same user database for all languages. For
example, when trigrams are added to the user database while typing
English, they do not really hurt when typing German later. If the
previous two words are German, English trigrams will just not
match. But this makes it possible to switch automatically to typing
English while staying with the German engine. Just typing a few
English words while still using the German engines makes the English
trigrams match again.  So this gives automatic language switching for
the user trigrams for all engines using Latin script.