Add feature to display input method description to setup tool
Resolves: rhbz#1001581 - Help text not available

Also Remove the options “m17n_mim_name” and “other_ime” from the .conf files

This is too complicated and not needed. Having the option “imes”
is enough because a list of input methods can be specified
in that option. The first input method in that list can be the
default, i.e. the extra option “m17n_mim_name” is not needed
to specify the default. And it is not needed to set “other_ime = False”
and specify an input method in “m17n_mim_name” to limit the choice
to one input method either, it is enough to have only one entry
in the list of input methods in the “imes” option.