Add python-enchant support
pyhunspell is not available for Python3. python-enchant is available
both for Python2 and for Python3.

So we need python-enchant because of the Python3 port anyway.

And python-enchant actually also seems to work slightly better then
pyhunspell, it takes Unicode as input and returns Unicode whereas
pyhunspell needs its input in the encoding of the hunspell dictionary.
So in case of pyhunspell, ibus-typing-booster has to do some extra
encoding conversion.