Commit when hitting the borders of the preëdit with the arrow-keys
Resolves: rhbz#1140502 - Commit preedit even if someone press right-left arrow
See: https://bugzilla.redhat.com/show_bug.cgi?id=1140502

Just stopping and ignoring further arrow key presses when the
left or the right end of the preëdit is reached is not so useful.
It seems more useful to commit the preëdit in that case and move
the cursor out of the preëdit.