Improve code to handle the left and right arrow keys
Control-arrow-left and control-arrow-right should also commit when the
end of the preëdit string is reached, otherwise this is inconsistent
with the behaviour of arrow-left and arrow-right, I think.

I.e. now the first Control-arrow-left moves the cursor to the left of the
preëdit, another Control-arrow-left commits and sends a Control-arrow-left
to the application.