Add an option to disable whether the digits 1-9 are used as selection keys
Using the digits 1-9 as keys to select a candidate to commit
or to delete the candidate makes it impossible to use these digits
as normal input. This make typing numbers more difficult, they
can only be typed while the preëdit is empty then. Therefore,
numbers can not use completion when they are needed as select keys.

But the F1-F9 keys are also always available as selection keys.

So it is not an absolute requirment to make the digits 1-9 behave as
selection keys as well. It may be convenient though because the
distance between the 1-9 keys and the fingers is shorter than the
distance between the F1-F9 keys and the fingers.

Therefore, I added this options so the user can decide what is more
important, having the 1-9 keys behave as selection keys or using
them to input numbers.